Refactor code into separate files
